moe实现原生输入法 next功能
